requirements.txt
fastapi
uvicorn
gtts
google-generativeai
requests
pydantic
python-dotenv
gtts
chemas